+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
 import './globals.css';
-import { ToastContainer } from '@src/components/Toast';
+import { ToastContainer } from '@src/components/providers/Toast';
 
 export const metadata: Metadata = {
   title: 'My App',
