+++ src/app/todo-list/page.tsx
@@ -1,4 +1,4 @@
-import TodoList from '@src/sections/TodoList';
+import TodoList from '@src/section/Todo';
 
 export default function TodoPage() {
   return <TodoList />;
